Add LuaConnector for autotracking N64 games
This commit adds a LuaConnector class which is capable of connecting to
emulator Lua scripts for autotracking. Proof of concept was done using
BizHawk 2.8 and the WarpWorld ConnectorLib lua script to power the
autotracking in my fork of Hamsda's ZOOTR pack.

This commit also adds an IAutotrackProvider interface which the
autotracker uses to abstract away the implementation of the
LuaConnector. Maybe the other trackers (snes, uat, ap) can be
migrated over to this interface to help clean up the autotracker code.